# openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Nexus REST API",
    "description": "Explore resources and get their data.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/v1/sources/registrations": {
      "get": {
        "tags": [
          "Sources"
        ],
        "summary": "Gets the list of data source registrations.",
        "operationId": "Sources_GetRegistrations",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "The optional user identifier. If not specified, the current user will be used.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/DataSourceRegistration"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sources"
        ],
        "summary": "Creates a data source registration.",
        "operationId": "Sources_CreateRegistration",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "The optional user identifier. If not specified, the current user will be used.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "registration",
          "description": "The registration to create.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataSourceRegistration"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/registrations/{registrationId}": {
      "delete": {
        "tags": [
          "Sources"
        ],
        "summary": "Deletes a data source registration.",
        "operationId": "Sources_DeleteRegistration",
        "parameters": [
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "description": "The identifier of the registration.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "userId",
            "in": "query",
            "description": "The optional user identifier. If not specified, the current user will be used.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataSourceRegistration": {
        "type": "object",
        "description": "A data source registration.",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the data source."
          },
          "resourceLocator": {
            "type": "string",
            "description": "An optional URL which points to the data.",
            "format": "uri",
            "nullable": true
          },
          "configuration": {
            "type": "object",
            "description": "Configuration parameters for the instantiated source.",
            "nullable": true,
            "additionalProperties": {}
          },
          "infoUrl": {
            "type": "string",
            "description": "An optional info URL.",
            "nullable": true
          },
          "releasePattern": {
            "type": "string",
            "description": "An optional regular expressions pattern to select the catalogs to be released. By default, all catalogs will be released.",
            "nullable": true
          },
          "visibilityPattern": {
            "type": "string",
            "description": "An optional regular expressions pattern to select the catalogs to be visible. By default, all catalogs will be visible.",
            "nullable": true
          }
        }
      },
...
    }
  }
}